Replace argument of writeProduceRequestV3
